move vstack test to PROTO_tests. Fix bug in vstack's use of np.common_type
